♻️ Refactor sidebar routes
